[R1464] adding quality badge to the term tree
